Trabalho_Faculdade/requirements.txt
python-dotenv==1.0.0
pywhatkit==5.4
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
pandas==2.0.3
nltk==3.8.1
feedparser==6.0.10
schedule==1.2.0
Flask==2.3.3
textblob==0.17.1
python-telegram-bot==13.15
twilio==8.5.0
matplotlib==3.7.2
scikit-learn==1.3.0 